Fix failure to login, fix deviation downloader script
deviantART has reworked their HTML a little meaning that all scripts failed
to log in, and then once that was fixed, deviation downloading failed completely
as the gallery widget had been completely redone. This has now been sorted, with
the exception that

*deviation folders are no longer savable*,

as the gallery widget no longer contains that information. It is available via a
separate pass to analyse the folders specially, but I don't need this currently
so it hasn't been implemented.